Add support for .env, fix strange sklearn issues
The former is done with python-dotenv. The latter is fixed by locking
sklearn, a depenency of openai, to 0.0.post1. The error:

Warning: The file chosen for install of sklearn 0.0.post2 (sklearn-0.0.post2.tar.gz) is yanked.
Reason for being yanked: Breaks poetry install that needs wheel building and does not have a fall-back to legacy install